Save as PDF functionality + Results Page Styling
Added a button to save the page as a PDF using html2pdf.js which seems to be the best solution. Saving the pdf via Chrome's print to PDF feature generates a cleaner PDF however there does not seem to be an API for this.

Also added majority of the styling for the results page, just need to update fonts and the download button.